Feat: 회원기능 에러메세지 구체화

+++ views/mypage/mypage.js
@@ -35,7 +35,10 @@
     
     try {
         const token = localStorage.getItem('token'); 
-        console.log(token)
+        if(!token){
+          alert("로그인 후 이용하세요!")
+          return
+        }
         const response = await fetch('http://localhost:5001/api/user', { 
             method: 'PUT',
             headers: {
@@ -58,7 +61,7 @@
             window.location.href = '/'; 
         } else {
             const errorData = await response.json();
-            console.error("Error updating user:", errorData);
+            alert("회원정보 수정오류: "+ errorData.error);
         }
     } catch (error) {
         console.error("Error:", error);
@@ -77,6 +80,10 @@
   
     
     const token = localStorage.getItem('token');
+    if(!token){
+      alert("로그인 후 이용하세요!")
+      return
+    }
     
     try {
         const response = await fetch('/api/order/user/orders',{
@@ -92,6 +99,7 @@
         displayOrders(data);
     } catch (error) {
         console.error('Error fetching orders:', error);
+        alert("주문내역이 없습니다!")
     }
   });
   
@@ -169,13 +177,16 @@
 const deleteAccountButton = document.getElementById('deleteAccountBtn');
 
   deleteAccountButton.addEventListener('click', async function() {
+    const token = localStorage.getItem('token');
+    if(!token){
+      alert("로그인 후 이용하세요!")
+      return
+    }
       // 확인 팝업창 띄우기
       const isConfirmed = window.confirm("정말로 탈퇴하시겠습니까?");
 
       if (isConfirmed) {
           try {
-              const token = localStorage.getItem('token');
-              
               const response = await fetch('http://localhost:5001/api/user', {
                   method: 'DELETE',
                   headers: {
@@ -207,12 +218,15 @@ document.addEventListener('DOMContentLoaded', (event) => {
   
   const logoutButton = document.getElementById('logoutButton');
   logoutButton.addEventListener('click', () => {
-    
+    const token = localStorage.getItem('token');
+    if(!token){
+      alert("로그인 후 이용하세요!")
+      return
+    }
     localStorage.removeItem('token');
     alert("로그아웃 하셨습니다.")
     // 메인 페이지로 리다이렉션
     window.location.href = '/'; 
-    console.log("로그아웃버튼이 눌러졌습니다!");
   });
 });
